Website Update
Completed steps 1 - 12 by adding the user login form to enter the user's username and password, adding the payment form for the user to enter billing information, and creating the image buttons to select which type of credit card. Also made a data list for the state input field to display all 50 state abbreviations.